+++ examples/10-openapi/src/app/modules/routed/first/first.module.ts
@@ -1,13 +1,8 @@
-import { Module, ModuleWithParams } from '@ditsmod/core';
-import { OpenapiModule } from '@ditsmod/openapi';
+import { Module } from '@ditsmod/core';
 
-import { openapiModuleWithParams } from '../../services/openapi/openapi.module';
 import { FirstController } from './first.controller';
 
-const firstOpenapiModule: ModuleWithParams<OpenapiModule> = { ...openapiModuleWithParams, path: '' };
-
 @Module({
-  imports: [firstOpenapiModule],
   controllers: [FirstController],
 })
 export class FirstModule {}