+++ packages/router/src/index.ts
@@ -2,3 +2,4 @@ export { DefaultRouter } from './router';
 export { Tree } from './tree';
 export { RouteParam } from './types';
 export { RouterModule } from './router.module';
+export { RouterLogMediator } from './router-log-mediator';