+++ packages/router/src/tree.ts
@@ -1,5 +1,6 @@
 import { Injectable, Optional } from '@ts-stack/di';
 
+import { RouterLogMediator } from './router-log-mediator';
 import { Fn, TreeConfig, RouteType, RouteParam } from './types';
 
 @Injectable()
@@ -12,7 +13,7 @@ export class Tree {
   protected indices: string;
   protected priority: number;
 
-  constructor(@Optional() treeConfig?: TreeConfig) {
+  constructor(private log: RouterLogMediator, @Optional() treeConfig?: TreeConfig) {
     Object.assign(this, new TreeConfig(), treeConfig);
   }
 
@@ -100,7 +101,7 @@ export class Tree {
     } else if (i == path.length) {
       // Make node a (in-path leaf)
       if (tree.handle !== null) {
-        throw new Error(`A handle is already registered for path '${fullPath}'`);
+        this.log.throwHandleAlreadyRegistered(fullPath);
       }
       tree.handle = handle;
     }
@@ -122,8 +123,7 @@ export class Tree {
       let end = i + 1;
       while (end < max && path[end] != '/') {
         if (path[end] == ':' || path[end] == '*') {
-          const msg = `only one wildcard per path segment is allowed, has: '${path.slice(i)}' in path '${fullPath}'`;
-          throw new Error(msg);
+          this.log.throwOnlyOneWildcardPerPath(path.slice(i), fullPath);
         } else {
           end++;
         }
@@ -132,14 +132,12 @@ export class Tree {
       // Check if this Tree existing children which would be unreachable
       // if we insert the wildcard here
       if (tree.children.length > 0) {
-        throw new Error(
-          `wildcard route '${path.slice(i, end)}' conflicts with existing children in path '${fullPath}'`
-        );
+        this.log.throwWildcardRouteConflicts(path.slice(i, end), fullPath);
       }
 
       // check if the wildcard has a name
       if (end - i < 2) {
-        throw new Error(`wildcards must be named with a non-empty name in path '${fullPath}'`);
+        this.log.throwWildcardsMustNonEmpty(fullPath);
       }
 
       if (c == ':') {
@@ -175,16 +173,16 @@ export class Tree {
         }
       } else {
         if (end != max || numParams > 1) {
-          throw new Error(`catch-all routes are only allowed at the end of the path in path '${fullPath}'`);
+          this.log.throwCatchAllRoutesOnlyAtEnd(fullPath);
         }
 
         if (tree.path.length > 0 && tree.path[tree.path.length - 1] == '/') {
-          throw new Error(`catch-all conflicts with existing handle for the path segment root in path '${fullPath}'`);
+          this.log.throwCatchAllConflictWithExistingHandle(fullPath);
         }
 
         i--;
         if (path[i] != '/') {
-          throw new Error(`no / before catch-all in path '${fullPath}'`);
+          this.log.throwNoBeforeCatchAll(fullPath);
         }
 
         tree.path = path.slice(offset, i);
@@ -245,9 +243,7 @@ export class Tree {
       pathSeg = path.split('/')[0];
     }
     const prefix = fullPath.slice(0, fullPath.indexOf(pathSeg)) + tree.path;
-    throw new Error(
-      `'${pathSeg}' in new path '${fullPath}' conflicts with existing wildcard '${tree.path}' in existing prefix '${prefix}'`
-    );
+    this.log.throwConflictsWithExistingWildcard(pathSeg, fullPath, tree.path, prefix);
   }
 
   protected splitAdge(tree: this, path: string, i: number) {
@@ -333,7 +329,7 @@ export class Tree {
               return { handle, params };
 
             default:
-              throw new Error('invalid node type');
+              this.log.throwInvalidNodeType();
           }
         }
       } else if (path == tree.path) {
@@ -345,7 +341,7 @@ export class Tree {
   }
 
   protected newTree(treeConfig?: TreeConfig) {
-    return new (this.constructor as typeof Tree)(treeConfig) as this;
+    return new (this.constructor as typeof Tree)(this.log, treeConfig) as this;
   }
 
   protected addPriority(pos: number) {